+++ python-threatexchange/threatexchange/extensions/video_vpdq/tests/test_vpdq_hash_and_match.py
@@ -1,6 +1,5 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 import pytest
-import vpdq_test_util
 from pathlib import Path
 import sys
 
@@ -13,7 +12,12 @@ except (ImportError, ModuleNotFoundError) as e:
     _DISABLED = True
 else:
     from threatexchange.extensions.video_vpdq.video_vpdq import VideoVPDQSignal
-    from vpdq_util import vpdq_to_json, TARGET_MATCH_PERCENT, QUERY_MATCH_PERCENT
+    from vpdq_util import (
+        vpdq_to_json,
+        read_file_to_hash,
+        TARGET_MATCH_PERCENT,
+        QUERY_MATCH_PERCENT,
+    )
 
 VIDEO = "test_video.mp4"
 HASH = "test_hash.txt"
@@ -24,7 +28,7 @@ WORKDIR = Path.cwd()
 class TestVPDQHasherMatcher:
     def test_vpdq_from_string_path(self):
         computed_hash = VideoVPDQSignal.hash_from_str(str(WORKDIR / VIDEO))
-        expected_hash = vpdq_test_util.read_file_to_hash(str(WORKDIR / HASH))
+        expected_hash = read_file_to_hash(str(WORKDIR / HASH))
         res, _ = VideoVPDQSignal.compare_hash(
             computed_hash, vpdq_to_json(expected_hash)
         )
