+++ python-threatexchange/threatexchange/extensions/video_vpdq/vpdq_util.py
@@ -12,7 +12,7 @@ QUERY_MATCH_PERCENT = "query_match_percent"
 
 def vpdq_to_json(vpdq_features: t.List[vpdq.VpdqFeature]) -> str:
     """Convert from VPDQ features to json object and return the json object as a str"""
-    data = {}  # data: dict
+    data = {}  # data: Dict[str, Dict] = ...
     for feature in vpdq_features:
         frame_number = feature.frame_number
         data[frame_number] = {}
@@ -67,3 +67,27 @@ def quality_filter(
         list of VPDQ feature: List of VPDQeatures with quality higher than distance_tolerance
     """
     return list(filter(lambda hash: hash.quality >= quality_tolerance, hashes))
+
+
+def read_file_to_hash(input_hash_filename: str) -> t.List[vpdq.VpdqFeature]:
+    """Read hash file and return vpdq hash
+
+    Args:
+        input_hash_filename (str): Input hash file path
+
+    Returns:
+        list of VpdqFeature: vpdq hash from the hash file"""
+
+    hash = []
+    with open(input_hash_filename, "r") as file:
+        lines = file.readlines()
+        for line in lines:
+            line = line.strip()
+            content = line.split(",")
+            pdq_hash = vpdq.str_to_hash(content[2])
+            feature = vpdq.VpdqFeature(
+                int(content[1]), int(content[0]), pdq_hash, float(content[3])
+            )
+            hash.append(feature)
+
+    return hash
